+++ UI/Controls/Controls.vb
@@ -463,12 +463,19 @@ Namespace UI
             If DesignHelp.IsDesignMode Then Exit Sub
 
             Dim cms As New ContextMenuStripEx()
+
             Dim cutItem = cms.Add("Cut")
             cutItem.SetImage(Symbol.Cut)
+            cutItem.ShortcutKeyDisplayString = "Ctrl+X"
+
             Dim copyItem = cms.Add("Copy", Sub() Clipboard.SetText(SelectedText))
             copyItem.SetImage(Symbol.Copy)
+            copyItem.ShortcutKeyDisplayString = "Ctrl+C"
+
             Dim pasteItem = cms.Add("Paste")
             pasteItem.SetImage(Symbol.Paste)
+            pasteItem.ShortcutKeyDisplayString = "Ctrl+V"
+
             cms.Add("Copy Everything", Sub() Clipboard.SetText(Text))
 
             AddHandler cutItem.Click, Sub()